關鍵字收尋 && 訂單CRUD && 串接金流

+++ migrations/20191209104353-create-order.js
@@ -21,7 +21,7 @@ module.exports = {
         type: Sequelize.STRING
       },
       total_amount: {
-        type: Sequelize.FLOAT
+        type: Sequelize.INTEGER
       },
       name: {
         type: Sequelize.STRING
